+++ src/components/Layout/index.js
@@ -1,3 +1,3 @@
-export * from './Content'
+export * from './Columns'
 export * from './LandingPage'
 export * from './Page'
